moved openvpn/client-server and their configs to node
