+++ core/connection/openvpn.go
@@ -29,6 +29,7 @@ import (
 	openvpn_bytescount "github.com/mysteriumnetwork/node/openvpn/middlewares/client/bytescount"
 	"github.com/mysteriumnetwork/node/openvpn/middlewares/state"
 	"github.com/mysteriumnetwork/node/server"
+	openvpn_node "github.com/mysteriumnetwork/node/services/openvpn"
 	"github.com/mysteriumnetwork/node/services/openvpn/middlewares/client/bytescount"
 	openvpn_session "github.com/mysteriumnetwork/node/services/openvpn/session"
 	"github.com/mysteriumnetwork/node/session"
@@ -43,13 +44,13 @@ func ConfigureVpnClientFactory(
 	originalLocationCache location.Cache,
 ) VpnClientCreator {
 	return func(vpnSession session.SessionDto, consumerID identity.Identity, providerID identity.Identity, stateCallback state.Callback, options ConnectOptions) (openvpn.Process, error) {
-		var receivedConfig openvpn.VPNConfig
+		var receivedConfig openvpn_node.VPNConfig
 		err := json.Unmarshal(vpnSession.Config, &receivedConfig)
 		if err != nil {
 			return nil, err
 		}
 
-		vpnClientConfig, err := openvpn.NewClientConfigFromSession(&receivedConfig, configDirectory, runtimeDirectory)
+		vpnClientConfig, err := openvpn_node.NewClientConfigFromSession(&receivedConfig, configDirectory, runtimeDirectory)
 		if err != nil {
 			return nil, err
 		}
@@ -70,7 +71,7 @@ func ConfigureVpnClientFactory(
 
 		credentialsProvider := openvpn_session.SignatureCredentialsProvider(vpnSession.ID, signer)
 
-		return openvpn.NewClient(
+		return openvpn_node.NewClient(
 			openvpnBinary,
 			vpnClientConfig,
 			state.NewMiddleware(stateCallback, statsSender.StateHandler),
