+++ services/openvpn/client.go
@@ -18,13 +18,14 @@
 package openvpn
 
 import (
+	"github.com/mysteriumnetwork/node/openvpn"
 	"github.com/mysteriumnetwork/node/openvpn/management"
 	"github.com/mysteriumnetwork/node/openvpn/tls"
 )
 
 // NewClient creates openvpn client with given config params
-func NewClient(openvpnBinary string, config *ClientConfig, middlewares ...management.Middleware) Process {
-	return CreateNewProcess(openvpnBinary, config.GenericConfig, middlewares...)
+func NewClient(openvpnBinary string, config *ClientConfig, middlewares ...management.Middleware) openvpn.Process {
+	return openvpn.CreateNewProcess(openvpnBinary, config.GenericConfig, middlewares...)
 }
 
 //VPNConfig structure represents VPN configuration options for given session
